+++ lib/logstash/json.rb
@@ -32,15 +32,23 @@ module LogStash
     ### JRuby
 
     def jruby_load(data, options = {})
-      options[:symbolize_keys] ? JrJackson::Raw.parse_sym(data) : JrJackson::Raw.parse_raw(data)
+      # TODO [guyboertje] remove these comments in 5.0
+      # options[:symbolize_keys] ? JrJackson::Raw.parse_sym(data) : JrJackson::Raw.parse_raw(data)
+
+      JrJackson::Ruby.parse(data, options)
+
     rescue JrJackson::ParseError => e
       raise LogStash::Json::ParserError.new(e.message)
     end
 
     def jruby_dump(o)
+      # TODO [guyboertje] remove these comments in 5.0
       # test for enumerable here to work around an omission in JrJackson::Json.dump to
       # also look for Java::JavaUtil::ArrayList, see TODO submit issue
-      o.is_a?(Enumerable) ? JrJackson::Raw.generate(o) : JrJackson::Json.dump(o)
+      # o.is_a?(Enumerable) ? JrJackson::Raw.generate(o) : JrJackson::Json.dump(o)
+
+      JrJackson::Base.generate(o, {})
+
     rescue => e
       raise LogStash::Json::GeneratorError.new(e.message)
     end